Fixed ownership of 2018Offseason directory
The 2018Offseason directory in the current Docker image is owned by root:root instead of ubuntu:ubuntu. To avoid respinning the Docker image, I added a command to fix the ownership. The current docker-run command also mounts the 2018Offseason repository as a volume from the user starting the container so it may be necessary to leave the permissions as-is. I'll do some testing later and see if that is needed or not.