Skills dashboard (#34)
* Add sample path images for robot paths

* Update Dashboard UI file

Also adds the AtHome dashboard, which is slightly different

* Refactor part of dashboard __init__

* Add sample file for drawable area

* Fix Dashboard UI

The dashboard removed some parts from the UI, which caused some code
crashes. Re-adding the widgets fixes the issue

* Fix new dashboard to have range displays

* Add Draw pad to main widget

This commit fixes the issue where the scribble area wasn't being added
to the dashboard properly. We had to remove the placeholder widget from
the original UI and instead just create a new widget to add to the
layout dynamically.

* Add teleop checkbox to UI and rename objects

* Add barrel racing/slalom paths to resources

* Load image based on combobox and resize dashboard

This commit will load the known path images for the dashboard and change
them based on the combobox selection.

* Clear drawing area when teleop box unchecked

The reason we do this is because in auto mode, we don't want any user
drawings distracting from the main map. In auto, your drawn path doesn't
do anything, so just leave it out.

Dashboard defaults to auto. When the teleop box is checked, users can
draw again.

* Allow different maps to be used for paths

Instead of hard-coding the paths, the map images can now be replaced
inside the "maps/" directory. It'll still have to be called something
specific. We'll add some config file later on for this

* Add dummy code for "execute path" and clear button

Clear the drawing area when the clear button is pressed.
When execute path is selected, just print the list of coordinates for
now. We'll eventually transform these coordinates for sending on the
actual robot

* Delete YAML map configuration

* Make map images full configurable

All map files must be placed in `resource/maps`, but otherwise, any file
can be specified for any path. The code updates support this change.

* Add constants for in/out of range images

* Print more detailed log information

* Add constants for background stles

* Add constants map for Ball images

* Remove unused functions in scribble.py

* Only draw on non-black parts of map

This helps prevent someone from driving a path that leads directly past
the map boundaries.

* Convert loaded images to RGB32

* Display robot position from `predicted_pose`

This commit will show the current robot position when a valid map is
loaded. Additionally, if the user clicks on the map, it will draw from
the path from the robot position.

* Publish `enable_in_teleop` based on checkbox

This commit adds a checkbox to the main dashboard UI. We also begin
publish the `/enable_in_teleop` topic based on the status of the
checkbox. This value is used in the auto code to determine if the robot
is okay to run auto code

* Return world coordinates when executing path

Previously, we returned coordinates relative to the top-left corner of
the screen, in pixels. We now return coordinates of the points selected
in real-world coordinates (in meters)

* Generate spline by running base_trajectory call

This commit allows the dashboard to generate the spline for the drawn
path on the dashboard. We currently get the generated spline and will
need to pass this along once the main branch is updated with other
changes

* Show generated spline as an overlay on draw pad

* Load path plans based on map_cfg.json

Instead of hard-coding the maps/paths to load, just load them all after
the map configuration is loaded. The information is in there anyway and
that will make everything more consistent

* Add transparency to generated robot path

* Remove files added by accident. Oops...

* Update predicted_pose namespace for subscriber

* Update enable in teleop topic publisher name

* Revert robot interface modification

* Remove unused maps from dashboard resource file

* Add service call for running the dynamic path

The actual service call is commented out since it's not available, but
the interface should be correct.

* added rest of field maps

Co-authored-by: Yashas Ambati <yashas.amb@gmail.com>