Ctre sim support (#28)
* Replace Talon FX interfaces with Talon SRX in sim

This commit changes the Talon FX motor controllers to use
the Talon SRX interfaces in order to support the CTRE sim stuff,
since there is no support for Talon FX sim right now.

In order to do this replacement, we stop the old read threads with the
FX interfaces, modify the shared pointers for the FX interfaces to use
SRX and then restart the thread. Note, this is all done in the sim init
code so it shouldn't affect the real robot.

* Generate encoder values based on mode

In position mode: pos = setpoint, vel = 0;
In velocity mode: pos = pos + vel*dt, vel = setpoint;
In motion magic:  pos = getActiveTrajPos(), vel = getActiveTrajVel();

* Fix typos and namespace errors

* Fix Sim update code to use `GetSimCollection()`

Also fixes a typo for checking if thread is joinable

Majorly, we updated the update code with the following:
* Determine conversion factors since we need to set stuff in radians
* Use the talon_command_ to set encoder feedback since talon_state_
won't actually update anything
* Use `SetQuadrature<X>` to update sim encoders

* Add debug messages

* Assorted cleanup to remove workarounds now that there's CTRE
sim support for TalonFX motors.

* More work to hook up to the various different versions of per-motor-type SimCollections

* Minor clean-up, tested working with motion magic at least

Co-authored-by: Vignesh Dhanasekaran <v.dhanasekaran0@gmail.com>