Auto note pickup (#114)
* add notes to stage map

* Fix auto node bug with premade paths

waypointIdxs wasn't being filled in, so when the path follower tried to
access that vector it segfaulted

* start thinking about how to auto-pickup notes

* autonomously driving to notes works pretty well in sim

* improve and simplify auto note pickup

* whole sim map addition

* align to object now starts at current velocity

* Rename to drive to object & change cmd_vel_mux

Note: drive_to_object now has a higher priority than teleop. This is
probably a bad idea. However, when it was at a lower priority than
teleop, it had to wait for teleop to time out before taking control
(which forced the robot to stop). I didn't want to raise the timeout in
case we drop messages.

We need to think of a better way to do this though.

* switch back to normal cmd vel timeout

* fix priority & change timeout

* cleanup after rebase

* 9000 camera offset

* change 9000 bot front

* WORKS IN REAL LIFE!!

* start of adding auto drive to notes in auto

* test

* lots of fixes and working auto drive to notes ! !

* cleanup

* changes from review, also rebase

---------

Co-authored-by: cienors <nathanlibutan37@gmail.com>
Co-authored-by: Torchtopher <torchtopher@gmail.com>