Add optional setting for controlling velocity ramp
This commit adds the variable 'traction_gain_percent_' in talon_swerve_drive_controller, which allows for limiting of velocity increases on the swerve drive.
It controls the percent increase between the current velocity and the goal velocity. If the goal velocity is too high, it is capped. This should hopefully help with traction issues.
If the variable is not set, the goal velocity is sent to the wheels as normal.