Clean up talon->cancoder sim interactions (#212)
* Clean up talon->cancoder sim interactions

Expose RobotHW pointer to Devices as a static member. Allow derived
classes to access it, getting interfaces and handles as if each devices
class was a controller.
Add a cancoder sim command interface allowing sets of various sim
fields.
Get cancoder sim command interface inside talon simRead() and use it to
query cancoder config and write sim values

Need to hook up sim cancoder device update() function which actually
writes the queued commands to the CTRE cancoder sim

* Hook up cancoder simRead (really simUpdate, but w/e)

* Testing talonfxpro sim with cancoder feedback

* Remove debugging printout

* Add talonfx sim command interface

Goal is to break up the queuing of new sim state and the actual writing
of it to the CTRE libraries. This mirrors the idea of having a
read/update/write loop for controllers, but in this case the read gets
the results of the CTRE sim, update queues up the new position/velocity
and write updates the CTRE sim collection values.

Should make it easier to keep the order of operations correct and
prevent various sim code from stepping on each other.

* More updates for sim code - seems to be working but seeing a weird pi/2 shift in cancoder values?

* Remove debugging printout

* Use forward declaration instead of including full .h file