Try to do it the right way
Found the bug leading to zero velocity: setAddRotorPosition (in TalonFX
command interface, and a similar one in CANcoder) was only calling the
hardware AddRotorPosition method if the new position to add was
different fron the old one. This has problems if you're moving at a
constant velocity :( (it's fixed now)

Also fixed the ordering of things, but one random swerve angle motor
continues to start with a very high velocity. I think this could be
related to the read thread running in the background (if you use a
TalonFX pointer, there's no delay between queueing and sending sim
commands, but there is a delay using the sim command interface, and also
the issue seemed to pop up more if I enabled before all read threads
were started even if I used the TalonFX pointer). Not exactly sure the
right way to do this (or even what exactly is happening, maybe it's
because we're writing to state_ in simRead of sim_talonfxpro_device
which is later overwritten when we copy from the read thread? idk), but
some comments are in flywheel_simulator.h