Allow outside nodes to force teleop node to drive rotation (#33)
* Allow outside nodes to force teleop node to drive rotation

Add a field to the teleop_orientation message to enable the teleop node
to rotate the robot even if joystick input is 0s.  This is useful in
cases where an outside node wants to drive the robot to a specific
orientation but doesn't need to overide the x&y cmd_vel fields by
publishing its own Twist message.  Should fix the case where the 2025
auto-rotate code doesn't rotate the robot unless it is also strafing
from joystick commannds.

* Rebase and clean up teleop logic a bit

* Add include lost during rebase